Tested additive variant of model C.
Tested a variant of model C with additive effects of both markers (rather than just marker 2). While this is the simplest way to improve the performance of model C while also leaving the performance metrics of models A and B unchanged, the text is explicit that model C should not include marker 1 (e.g., Figure 3), making a typo unlikely. Regardless, the results don't agree with Table 1.